Move mock LLM provider to internal/llm/llmtest
Keeps MockProvider out of the production binary — the llmtest package
is only imported by test code. Package name is llmtest so callers use
llmtest.MockProvider rather than llm.MockProvider, making the test
boundary explicit at the import site.
